+++ sdk/cosmos/azure_data_cosmos_driver/src/models/mod.rs
@@ -53,7 +53,7 @@ pub use cosmos_headers::{
     AutoscaleAutoUpgradePolicy, AutoscaleThroughputPolicy, CosmosRequestHeaders,
     CosmosResponseHeaders, MaxItemCountHint, OfferAutoscaleSettings,
 };
-pub use cosmos_operation::CosmosOperation;
+pub use cosmos_operation::{ChangeFeedStartFrom, CosmosOperation};
 pub use cosmos_resource_reference::CosmosResourceReference;
 pub(crate) use cosmos_resource_reference::ResourcePaths;
 pub use cosmos_response::CosmosResponse;